Review feedback addressed:
 * decrease the risk that the new factory is used from the wrong places
 * fallback to the old inline impl on 2.12 when reflection isn't possible
